+++ app/(public)/page.tsx
@@ -1,7 +1,9 @@
 'use client'
 import { Flex } from '@chakra-ui/react'
-
+import { useSession } from 'next-auth/react'
 export default function Home() {
+  const { data: session, status } = useSession()
+  console.log(session)
   return (
     <Flex height='100vh' alignItems='center' justifyContent='center'>
       <>
